+++ R/print.randomPhiDivTest.r
@@ -1,0 +1,10 @@
+print.randomPhiDivTest <- function(x){
+  cat("\n")
+  cat("\t Goodness-of-fit test using Random phi-divergence statistics: \n")
+  cat("\n")
+  cat("data")
+
+
+
+  invisible(x)
+}