+++ NEWS.md
@@ -1,3 +1,11 @@
+# RPDTest 0.0.2
+
+- Bug fixes in calculating statistics
+- New method for simulate p-value
+- Update descriptions
+
+---
+
 # RPDTest 0.0.1
 
 * Initial CRAN submission.